Avoid using deprecated Buffer API on newer Node.js
This avoids using Buffer constructor API on newer Node.js versions.

To achieve that, Buffer.from presence is checked, with validation that it's
not the same method as Uint8Array.from.

Also an additional type-guard is added in the fallback code path to ensure
that typed numbers are never accidently fed into `new Buffer` input.

Refs:
https://nodejs.org/api/deprecations.html#deprecations_dep0005_buffer_constructor