add support for mixing the keyId and algorithm into the message to sign
this has been implemented locally in both a perl and c version of
this scheme, and seems to be working well. i'm adding it here to
help interoperability, and to make it easier to test the other
codebases against this version, which i consider the canonical
implementation.

discussed with alex wilson