refactor generation of the authorization header value.
if options.headers is not specified, do not include it in the params.
this is handled by the parser side of things because if headers it
not specified, everything assumes the date (or x-date) header will
be signed.

this also paves the way to include more params in the authorization
header.