add support for an opaque param in the authorization header.
the idea and semantics are largely copied from http digest auth,
and is nicely described in rfc2617 as:

>   opaque
>     A string of data, specified by the server, which should be returned
>     by the client unchanged in the Authorization header of subsequent
>     requests with URIs in the same protection space. It is recommended
>     that this string be base64 or hexadecimal data.

that rfc is obsoleted by rfc7616, which doesn't really change the
semantics and use of opaque, it just takes more words to describe
it.

the idea is to mitigate the reuse of signature data between services,
or "protection spaces" on the same service. if a 3rd party could
capture the authorization header, we want to mitigate against them
replaying it against another service to authenticate as the user.

the current best practice for that is to require that at least the
"host" header and "(request-target)" pseudo-header (and maybe the
"(keyid)") and have them included included as input for the signature
generation. this is fine, but this means the client cannot reuse a
signature and will have to generate one per request. if they're
using something like a yubikey to sign requests, and they're doing
a lot of requests, this feels slow.

using an "opaque" value can mitigate this. letting the server provide
an "opaque" value as part of a WWW-Authenticate rely header means
it can scope a "protection space" with an identifier it provides.
you can then use "(opaque)" instead of "(request-target)" as an
input to the signature. if servers provide unique opaque values,
they can mitigate against reuse of the signature between services,
while allowing a client to reuse a signature across multiple requests
to the same service.

discussed with alex wilson.